+++ CHANGELOG.md
@@ -5,3 +5,5 @@
 ### Changed
 
 - Change MaxDataServiceVersion to 3.0 in requests.
+- Change DataServiceVersion to 2.0 in requests.
+- Made `$type` argument optional in `Mekras\OData\Client::createEntityDocument()`.